requirements.txt
discord.py==0.16.3
aiohttp
websockets
chardet
cassiopeia